Add a method for converting structs -> maps when bridging MM -> TF
