+++ google/resource_google_folder.go
@@ -63,8 +63,12 @@ func resourceGoogleFolderCreate(d *schema.ResourceData, meta interface{}) error 
 		return fmt.Errorf("Error creating folder '%s' in '%s': %s", displayName, parent, err)
 	}
 
-	err = resourceManagerV2Beta1OperationWait(config.clientResourceManager, op, "creating folder")
+	opAsMap, err := ConvertToMap(op)
+	if err != nil {
+		return err
+	}
 
+	err = resourceManagerOperationWaitTime(config, opAsMap, "", "creating folder", int(d.Timeout(schema.TimeoutCreate).Minutes()))
 	if err != nil {
 		return fmt.Errorf("Error creating folder '%s' in '%s': %s", displayName, parent, err)
 	}
@@ -132,7 +136,12 @@ func resourceGoogleFolderUpdate(d *schema.ResourceData, meta interface{}) error 
 			return fmt.Errorf("Error moving folder '%s' to '%s': %s", displayName, newParent, err)
 		}
 
-		err = resourceManagerV2Beta1OperationWait(config.clientResourceManager, op, "move folder")
+		opAsMap, err := ConvertToMap(op)
+		if err != nil {
+			return err
+		}
+
+		err = resourceManagerOperationWaitTime(config, opAsMap, "", "move folder", int(d.Timeout(schema.TimeoutCreate).Minutes()))
 		if err != nil {
 			return fmt.Errorf("Error moving folder '%s' to '%s': %s", displayName, newParent, err)
 		}
